rearrange pt camera functions
