+++ src/materials/pathtracing/PhysicalPathTracingMaterial.js
@@ -38,6 +38,8 @@ import { arraySamplerTexelFetchGLSL } from '../../shader/common/arraySamplerTexe
 import { pcgGLSL } from '../../shader/rand/pcg.glsl.js';
 import { sobolCommonGLSL, sobolSamplingGLSL } from '../../shader/rand/sobol.glsl.js';
 
+// path tracer utils
+import { cameraUtilsGLSL } from './glsl/cameraUtils.glsl.js';
 
 export class PhysicalPathTracingMaterial extends MaterialBase {
 
@@ -151,45 +153,55 @@ export class PhysicalPathTracingMaterial extends MaterialBase {
 				${ shapeSamplingGLSL }
 				${ bsdfSamplingGLSL }
 				${ equirectSamplingGLSL }
+				${ lightSamplingGLSL }
 
+				// environment
+				uniform EquirectHdrInfo envMapInfo;
 				uniform mat4 environmentRotation;
+				uniform float environmentIntensity;
+
+				// lighting
+				uniform sampler2DArray iesProfiles;
+				uniform LightsInfo lights;
+
+				// background
 				uniform float backgroundBlur;
 				uniform float backgroundAlpha;
-
 				#if FEATURE_BACKGROUND_MAP
 
 				uniform sampler2D backgroundMap;
 
 				#endif
 
+				// camera
+				uniform mat4 cameraWorldMatrix;
+				uniform mat4 invProjectionMatrix;
 				#if FEATURE_DOF
 
 				uniform PhysicalCamera physicalCamera;
 
 				#endif
 
-				uniform vec2 resolution;
-				uniform int bounces;
-				uniform int transmissiveBounces;
-				uniform mat4 cameraWorldMatrix;
-				uniform mat4 invProjectionMatrix;
+				// geometry
 				uniform sampler2DArray attributesArray;
 				uniform usampler2D materialIndexAttribute;
+				uniform sampler2D materials;
+				uniform sampler2DArray textures;
 				uniform BVH bvh;
-				uniform float environmentIntensity;
+
+				// path tracer
+				uniform int bounces;
+				uniform int transmissiveBounces;
 				uniform float filterGlossyFactor;
 				uniform int seed;
+
+				// image
+				uniform vec2 resolution;
 				uniform float opacity;
-				uniform sampler2D materials;
-				uniform LightsInfo lights;
-				uniform sampler2DArray iesProfiles;
 
-				${ lightSamplingGLSL }
-
-				uniform EquirectHdrInfo envMapInfo;
-
-				uniform sampler2DArray textures;
 				varying vec2 vUv;
+
+				${ cameraUtilsGLSL }
 
 				float applyFilteredGlossy( float roughness, float accumulatedRoughness ) {
 
@@ -358,84 +370,6 @@ export class PhysicalPathTracingMaterial extends MaterialBase {
 					}
 
 					return true;
-
-				}
-
-				vec3 ndcToRayOrigin( vec2 coord ) {
-
-					vec4 rayOrigin4 = cameraWorldMatrix * invProjectionMatrix * vec4( coord, - 1.0, 1.0 );
-					return rayOrigin4.xyz / rayOrigin4.w;
-				}
-
-				void getCameraRay( out vec3 rayDirection, out vec3 rayOrigin ) {
-
-					vec2 ssd = vec2( 1.0 ) / resolution;
-
-					// Jitter the camera ray by finding a uv coordinate at a random sample
-					// around this pixel's UV coordinate for AA
-					vec2 ruv = sobol2( 0 );
-					vec2 jitteredUv = vUv + vec2( tentFilter( ruv.x ) * ssd.x, tentFilter( ruv.y ) * ssd.y );
-
-					#if CAMERA_TYPE == 2
-
-						// Equirectangular projection
-						vec4 rayDirection4 = vec4( equirectUvToDirection( jitteredUv ), 0.0 );
-						vec4 rayOrigin4 = vec4( 0.0, 0.0, 0.0, 1.0 );
-
-						rayDirection4 = cameraWorldMatrix * rayDirection4;
-						rayOrigin4 = cameraWorldMatrix * rayOrigin4;
-
-						rayDirection = normalize( rayDirection4.xyz );
-						rayOrigin = rayOrigin4.xyz / rayOrigin4.w;
-
-					#else
-
-						// get [- 1, 1] normalized device coordinates
-						vec2 ndc = 2.0 * jitteredUv - vec2( 1.0 );
-						rayOrigin = ndcToRayOrigin( ndc );
-
-						#if CAMERA_TYPE == 1
-
-							// Orthographic projection
-							rayDirection = ( cameraWorldMatrix * vec4( 0.0, 0.0, - 1.0, 0.0 ) ).xyz;
-							rayDirection = normalize( rayDirection );
-
-						#else
-
-							// Perspective projection
-							rayDirection = normalize( mat3(cameraWorldMatrix) * ( invProjectionMatrix * vec4( ndc, 0.0, 1.0 ) ).xyz );
-
-						#endif
-
-					#endif
-
-					#if FEATURE_DOF
-					{
-
-						// depth of field
-						vec3 focalPoint = rayOrigin + normalize( rayDirection ) * physicalCamera.focusDistance;
-
-						// get the aperture sample
-						// if blades === 0 then we assume a circle
-						vec3 shapeUVW= sobol3( 1 );
-						int blades = physicalCamera.apertureBlades;
-						float anamorphicRatio = physicalCamera.anamorphicRatio;
-						vec2 apertureSample = blades == 0 ? sampleCircle( shapeUVW.xy ) : sampleRegularPolygon( blades, shapeUVW );
-						apertureSample *= physicalCamera.bokehSize * 0.5 * 1e-3;
-
-						// rotate the aperture shape
-						apertureSample =
-							rotateVector( apertureSample, physicalCamera.apertureRotation ) *
-							saturate( vec2( anamorphicRatio, 1.0 / anamorphicRatio ) );
-
-						// create the new ray
-						rayOrigin += ( cameraWorldMatrix * vec4( apertureSample, 0.0, 0.0 ) ).xyz;
-						rayDirection = focalPoint - rayOrigin;
-
-					}
-					#endif
-
-					rayDirection = normalize( rayDirection );
 
 				}
