+++ src/materials/pathtracing/glsl/cameraUtils.glsl.js
@@ -1,0 +1,81 @@
+export const cameraUtilsGLSL = /* glsl */`
+
+	vec3 ndcToRayOrigin( vec2 coord ) {
+
+		vec4 rayOrigin4 = cameraWorldMatrix * invProjectionMatrix * vec4( coord, - 1.0, 1.0 );
+		return rayOrigin4.xyz / rayOrigin4.w;
+	}
+
+	void getCameraRay( out vec3 rayDirection, out vec3 rayOrigin ) {
+
+		vec2 ssd = vec2( 1.0 ) / resolution;
+
+		// Jitter the camera ray by finding a uv coordinate at a random sample
+		// around this pixel's UV coordinate for AA
+		vec2 ruv = sobol2( 0 );
+		vec2 jitteredUv = vUv + vec2( tentFilter( ruv.x ) * ssd.x, tentFilter( ruv.y ) * ssd.y );
+
+		#if CAMERA_TYPE == 2
+
+			// Equirectangular projection
+			vec4 rayDirection4 = vec4( equirectUvToDirection( jitteredUv ), 0.0 );
+			vec4 rayOrigin4 = vec4( 0.0, 0.0, 0.0, 1.0 );
+
+			rayDirection4 = cameraWorldMatrix * rayDirection4;
+			rayOrigin4 = cameraWorldMatrix * rayOrigin4;
+
+			rayDirection = normalize( rayDirection4.xyz );
+			rayOrigin = rayOrigin4.xyz / rayOrigin4.w;
+
+		#else
+
+			// get [- 1, 1] normalized device coordinates
+			vec2 ndc = 2.0 * jitteredUv - vec2( 1.0 );
+			rayOrigin = ndcToRayOrigin( ndc );
+
+			#if CAMERA_TYPE == 1
+
+				// Orthographic projection
+				rayDirection = ( cameraWorldMatrix * vec4( 0.0, 0.0, - 1.0, 0.0 ) ).xyz;
+				rayDirection = normalize( rayDirection );
+
+			#else
+
+				// Perspective projection
+				rayDirection = normalize( mat3(cameraWorldMatrix) * ( invProjectionMatrix * vec4( ndc, 0.0, 1.0 ) ).xyz );
+
+			#endif
+
+		#endif
+
+		#if FEATURE_DOF
+		{
+
+			// depth of field
+			vec3 focalPoint = rayOrigin + normalize( rayDirection ) * physicalCamera.focusDistance;
+
+			// get the aperture sample
+			// if blades === 0 then we assume a circle
+			vec3 shapeUVW= sobol3( 1 );
+			int blades = physicalCamera.apertureBlades;
+			float anamorphicRatio = physicalCamera.anamorphicRatio;
+			vec2 apertureSample = blades == 0 ? sampleCircle( shapeUVW.xy ) : sampleRegularPolygon( blades, shapeUVW );
+			apertureSample *= physicalCamera.bokehSize * 0.5 * 1e-3;
+
+			// rotate the aperture shape
+			apertureSample =
+				rotateVector( apertureSample, physicalCamera.apertureRotation ) *
+				saturate( vec2( anamorphicRatio, 1.0 / anamorphicRatio ) );
+
+			// create the new ray
+			rayOrigin += ( cameraWorldMatrix * vec4( apertureSample, 0.0, 0.0 ) ).xyz;
+			rayDirection = focalPoint - rayOrigin;
+
+		}
+		#endif
+
+		rayDirection = normalize( rayDirection );
+
+	}
+
+`;
